LayoutTests/imported/w3c/web-platform-tests/fetch/data-urls/processing.any.worker-expected.txt


PASS Setup. 
PASS "data://test/,X" 
PASS "data://test:test/,X" 
PASS "data:,X" 
PASS "data:" 
PASS "data:text/html" 
PASS "data:text/html    ;charset=x   " 
PASS "data:," 
FAIL "data:,X#X" assert_array_equals: lengths differ, expected 1 got 3
PASS "data:,%FF" 
PASS "data:text/plain,X" 
PASS "data:text/plain ,X" 
PASS "data:text/plain%20,X" 
FAIL "data:text/plain\f,X" assert_equals: expected "text/plain%0c" but got "text/plain%0C"
FAIL "data:text/plain%0C,X" assert_equals: expected "text/plain%0c" but got "text/plain%0C"
FAIL "data:text/plain;,X" assert_equals: expected "text/plain" but got "text/plain;"
PASS "data:;x=x;charset=x,X" 
PASS "data:;x=x,X" 
PASS "data:text/plain;charset=windows-1252,%C2%B1" 
FAIL "data:text/plain;Charset=UTF-8,%C2%B1" assert_equals: expected "text/plain;charset=UTF-8" but got "text/plain;Charset=UTF-8"
PASS "data:image/gif,%C2%B1" 
FAIL "data:IMAGE/gif,%C2%B1" assert_equals: expected "image/gif" but got "IMAGE/gif"
FAIL "data:IMAGE/gif;hi=x,%C2%B1" assert_equals: expected "image/gif;hi=x" but got "IMAGE/gif;hi=x"
FAIL "data:IMAGE/gif;CHARSET=x,%C2%B1" assert_equals: expected "image/gif;charset=x" but got "IMAGE/gif;CHARSET=x"
PASS "data: ,%FF" 
PASS "data:%20,%FF" 
PASS "data:\f,%FF" 
PASS "data:%1F,%FF" 
PASS "data:\0,%FF" 
PASS "data:%00,%FF" 
PASS "data:text/html  ,X" 
PASS "data:text / html,X" 
PASS "data:†,X" 
FAIL "data:†/†,X" assert_equals: expected "%e2%80%a0/%e2%80%a0" but got "%E2%80%A0/%E2%80%A0"
PASS "data:X,X" 
PASS "data:image/png,X X" 
PASS "data:application/javascript,X X" 
PASS "data:application/xml,X X" 
PASS "data:text/javascript,X X" 
PASS "data:text/plain,X X" 
PASS "data:unknown/unknown,X X" 
FAIL "data:text/plain;a=\",\",X" assert_equals: expected "text/plain;a=\"\"" but got "text/plain;a=\""
PASS "data:text/plain;a=%2C,X" 
FAIL "data:;base64;base64,WA" assert_equals: expected "text/plain" but got "text/plain;base64"
FAIL "data:x/x;base64;base64,WA" assert_equals: expected "x/x" but got "x/x;base64"
FAIL "data:x/x;base64;charset=x,WA" assert_equals: expected "x/x;charset=x" but got "x/x;base64;charset=x"
FAIL "data:x/x;base64;charset=x;base64,WA" assert_equals: expected "x/x;charset=x" but got "x/x;base64;charset=x"
FAIL "data:x/x;base64;base64x,WA" assert_equals: expected "x/x" but got "x/x;base64;base64x"
PASS "data:;base64,W%20A" 
PASS "data:;base64,W%0CA" 
PASS "data:x;base64x,WA" 
PASS "data:x;base64;x,WA" 
PASS "data:x;base64=x,WA" 
FAIL "data:; base64,WA" assert_array_equals: lengths differ, expected 1 got 2
FAIL "data:;  base64,WA" assert_array_equals: lengths differ, expected 1 got 2
FAIL "data:  ;charset=x   ;  base64,WA" assert_array_equals: lengths differ, expected 1 got 2
FAIL "data:;base64;,WA" assert_equals: expected "text/plain" but got "text/plain;base64;"
FAIL "data:;base64 ,WA" assert_array_equals: lengths differ, expected 1 got 2
FAIL "data:;base64   ,WA" assert_array_equals: lengths differ, expected 1 got 2
FAIL "data:;base 64,WA" assert_equals: expected "text/plain" but got "text/plain;base 64"
PASS "data:;BASe64,WA" 
FAIL "data:;%62ase64,WA" assert_equals: expected "text/plain" but got "text/plain;%62ase64"
PASS "data:%3Bbase64,WA" 
PASS "data:;charset=x,X" 
FAIL "data:; charset=x,X" assert_equals: expected "text/plain;charset=x" but got "text/plain; charset=x"
FAIL "data:;charset =x,X" assert_equals: expected "text/plain" but got "text/plain;charset =x"
FAIL "data:;charset= x,X" assert_equals: expected "text/plain;charset=\" x\"" but got "text/plain;charset= x"
FAIL "data:;charset=,X" assert_equals: expected "text/plain" but got "text/plain;charset="
FAIL "data:;charset,X" assert_equals: expected "text/plain" but got "text/plain;charset"
FAIL "data:;charset=\"x\",X" assert_equals: expected "text/plain;charset=x" but got "text/plain;charset=\"x\""
FAIL "data:;CHARSET=\"X\",X" assert_equals: expected "text/plain;charset=X" but got "text/plain;CHARSET=\"X\""
